requirements.txt
streamlit
pandas
google-genai
scikit-learn
numpy
st-gsheets-connection